Move component into components folder
